+++ mods/BrewLAN/documentation/Changelog.md
@@ -267,6 +267,11 @@ All changes by Sean Wheeldon (Balthazar) unless otherwise stated.
   * Replaced the icon of the Pigeon.
   * Pigeon now adopts an aggressive pitch when attacking.
 
+  ### 2018-06-05
+
+  * Redesigned the Aeon energy storage.
+  * Set the collision model of the Aeon energy storage to be spherical.
+
 ## 2019-01-03|84 (FAF minor patch)
 
 * Created a function to add the FAF category CQUEMOV to function to mobile experimental mobile units.